pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "payloadguard"
version = "1.0.0"
description = "Scans a branch before merge and tells you exactly how badly it's going to hurt."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "DarkVader-PLG", email = "warorder10wao@gmail.com" }]
requires-python = ">=3.8"
keywords = ["git", "security", "code-review", "static-analysis", "devops"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "GitPython>=3.1.41",
    "PyYAML>=6.0",
    "PyJWT[crypto]>=2.8.0",
    "requests>=2.31.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "tree-sitter-java>=0.21.0",
]

[project.scripts]
payloadguard = "analyze:main"

[project.urls]
Homepage = "https://github.com/DarkVader-PLG/payload-consequence-analyser"
Repository = "https://github.com/DarkVader-PLG/payload-consequence-analyser"
"Bug Tracker" = "https://github.com/DarkVader-PLG/payload-consequence-analyser/issues"

[tool.setuptools]
py-modules = ["analyze", "structural_parser", "post_check_run"]
